Order resolving reset fix, new Order Resolving Controller
