+++ android-client/courier/src/main/java/ru/commandos/diner/delivery/controller/OrderResolvingController.java
@@ -1,0 +1,15 @@
+package ru.commandos.diner.delivery.controller;
+
+import java.util.HashMap;
+import java.util.Map;
+
+import ru.commandos.diner.delivery.model.Order;
+
+public class OrderResolvingController {
+
+    private final Map<Order, Boolean> orders = new HashMap<>();
+
+    public Map<Order, Boolean> getOrders() {
+        return orders;
+    }
+}
